Server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
asyncpg
python-dotenv
geoalchemy2
shapely
pydantic-settings
geopandas